@@ -1,70 +1,67 @@
 {
     "private": true,
     "scripts": {
-        "dev": "npm run development",
-        "development": "mix",
-        "watch": "mix watch",
-        "watch-poll": "mix watch -- --watch-options-poll=1000",
-        "hot": "mix watch --hot",
-        "prettier": "prettier --write .",
-        "prod": "npm run production",
-        "production": "mix --production"
+        "dev": "vite",
+        "build": "vite build",
+        "prettier": "prettier --write resources/js/**/*.{js,jsx,ts,tsx}"
     },
     "devDependencies": {
-        "@babel/preset-react": "^7.16.7",
-        "@emotion/react": "^11.9.0",
-        "@emotion/styled": "^11.8.1",
+        "@babel/preset-react": "^7.18.6",
+        "@emotion/react": "^11.9.3",
+        "@emotion/styled": "^11.9.3",
         "@popperjs/core": "^2.11.5",
         "@types/draft-js": "^0.11.9",
         "@types/draftjs-to-html": "^0.8.1",
         "@types/html-to-draftjs": "^1.4.0",
-        "@types/jest": "^27.5.1",
-        "@types/node": "^17.0.32",
-        "@types/react": "^18.0.9",
-        "@types/react-dom": "^18.0.3",
+        "@types/jest": "^28.1.6",
+        "@types/node": "^18.6.1",
+        "@types/react": "^18.0.15",
+        "@types/react-dom": "^18.0.6",
         "@types/react-redux": "^7.1.24",
+        "@vitejs/plugin-react": "^2.0.0",
         "axios": "^0.27",
-        "bootstrap": "^5.1.3",
-        "eslint-plugin-react-hooks": "^4.5.0",
-        "html-loader": "^3.1.0",
-        "laravel-mix": "^6.0.43",
+        "bootstrap": "^5.2.0",
+        "eslint-plugin-react-hooks": "^4.6.0",
+        "html-loader": "^4.1.0",
+        "laravel-vite-plugin": "^0.5.2",
         "lodash": "^4.17.21",
-        "postcss": "^8.4.13",
+        "postcss": "^8.4.14",
         "postcss-focus": "^5.0.1",
-        "prettier": "^2.6.2",
+        "prettier": "^2.7.1",
         "resolve-url-loader": "^5.0.0",
-        "sass": "^1.51.0",
-        "sass-loader": "^12.6.0",
-        "ts-loader": "^9.3.0"
+        "sass": "^1.54.0",
+        "sass-loader": "^13.0.2",
+        "ts-loader": "^9.3.1",
+        "vite": "^3.0.4",
+        "vite-tsconfig-paths": "^3.5.0"
     },
     "dependencies": {
-        "@babel/core": "^7.17.12",
+        "@babel/core": "^7.18.9",
         "@fontsource/roboto": "^4.5.7",
-        "@mui/icons-material": "^5.6.2",
-        "@mui/material": "^5.7.0",
-        "@mui/system": "^5.7.0",
-        "@reduxjs/toolkit": "^1.8.1",
+        "@mui/icons-material": "^5.8.4",
+        "@mui/material": "^5.9.2",
+        "@mui/system": "^5.9.2",
+        "@reduxjs/toolkit": "^1.8.3",
         "draft-js": "^0.11.7",
         "draftjs-to-html": "^0.9.1",
-        "eslint": "^8.15.0",
+        "eslint": "^8.20.0",
         "html-to-draftjs": "^1.5.0",
-        "immutable": "^4.0.0",
-        "laravel-mix-workbox": "^0.1.4",
-        "react": "^18.1.0",
-        "react-dom": "^18.1.0",
-        "react-image-crop": "^10.0.1",
-        "react-multi-carousel": "^2.8.0",
-        "react-redux": "^8.0.1",
+        "immutable": "^4.1.0",
+        "react": "^18.2.0",
+        "react-dom": "^18.2.0",
+        "react-image-crop": "^10.0.4",
+        "react-redux": "^8.0.2",
         "react-router-dom": "^6.3.0",
-        "typescript": "^4.6.4",
-        "webpack": "^5.72.1",
+        "swiper": "^8.3.2",
+        "typescript": "^4.7.4",
+        "webpack": "^5.74.0",
         "workbox-webpack-plugin": "^6.5.3"
     },
     "prettier": {
         "trailingComma": "es5",
         "tabWidth": 4,
-        "semi": true,
         "singleQuote": true,
-        "printWidth": 120
+        "printWidth": 120,
+        "semi": false
     }
 }
