@@ -1,3 +1,3 @@
-declare module '*.jpg';
-declare module '*.png';
-declare module '*.webp';
+declare module '*.jpg'
+declare module '*.png'
+declare module '*.webp'
