@@ -1,20 +1,20 @@
-import { createApi } from '@reduxjs/toolkit/query/react';
-import { fetchBaseQuery } from '@reduxjs/toolkit/dist/query/react';
-import { gameObj, userState, userInfoData } from '../types';
-import { RootState } from '../store';
+import { createApi } from '@reduxjs/toolkit/query/react'
+import { fetchBaseQuery } from '@reduxjs/toolkit/dist/query/react'
+import { gameObj, userState, userInfoData } from '../types'
+import { RootState } from '../store'
 
 export const portalApi = createApi({
     reducerPath: 'portalApi',
     baseQuery: fetchBaseQuery({
         baseUrl: 'https://portaleducacional.tec.br',
         prepareHeaders: (headers, { getState }) => {
-            const token = (getState() as RootState).user.token;
+            const token = (getState() as RootState).user.token
             if (token) {
-                headers.set('authorization', `Bearer ${token}`);
+                headers.set('authorization', `Bearer ${token}`)
             }
-            headers.set('content-type', 'application/json');
-            headers.set('accept', 'application/json');
-            return headers;
+            headers.set('content-type', 'application/json')
+            headers.set('accept', 'application/json')
+            return headers
         },
     }),
     endpoints: (builder) => ({
@@ -51,13 +51,13 @@ export const portalApi = createApi({
             query: (id) => `/api/conteudo/${id}`,
         }),
     }),
-});
+})
 
 export const {
     useCreateGameObjectMutation,
     useEditGameObjectMutation,
     useGetUserInfoQuery,
     useGetGameObjectByIdQuery,
-} = portalApi;
+} = portalApi
 
-export default portalApi;
+export default portalApi
