@@ -1,5 +1,5 @@
-import { EditorState, ContentState } from 'draft-js';
-import htmlToDraft from 'html-to-draftjs';
+import { EditorState, ContentState } from 'draft-js'
+import htmlToDraft from 'html-to-draftjs'
 
 const textToDraft = (raw: string): EditorState => {
     let collection = {
@@ -11,13 +11,13 @@ const textToDraft = (raw: string): EditorState => {
         '\\[/u\\]': '</ins>',
         '\\[s\\]': '<del>',
         '\\[/s\\]': '</del>',
-    };
-    for (const [key, value] of Object.entries(collection)) {
-        const regex = new RegExp(key, 'g');
-        raw = raw.replace(regex, value);
     }
-    const { contentBlocks, entityMap } = htmlToDraft(raw);
-    return EditorState.createWithContent(ContentState.createFromBlockArray(contentBlocks, entityMap));
-};
+    for (const [key, value] of Object.entries(collection)) {
+        const regex = new RegExp(key, 'g')
+        raw = raw.replace(regex, value)
+    }
+    const { contentBlocks, entityMap } = htmlToDraft(raw)
+    return EditorState.createWithContent(ContentState.createFromBlockArray(contentBlocks, entityMap))
+}
 
-export default textToDraft;
+export default textToDraft
